requirements.txt
Django==3.1.6
django-credit-cards==0.4.1
Pillow==8.1.1